+++ standingsrequests/models.py
@@ -400,7 +400,7 @@ class StandingRequest(AbstractStandingsRequest):
 
     objects = StandingRequestManager()
 
-    def delete(self, using=None, keep_parents=False):
+    def delete(self, using=None, keep_parents=False, reason=None):
         """
         Add a revocation before deleting if the standing has been
         actioned (pending) or is effective and
@@ -419,6 +419,7 @@ class StandingRequest(AbstractStandingsRequest):
                     contact_id=self.contact_id,
                     contact_type=self.contact_id_2_type(self.contact_type_id),
                     user=self.user,
+                    reason=reason,
                 )
             else:
                 logger.debug(
@@ -508,6 +509,7 @@ class StandingRevocation(AbstractStandingsRequest):
         OWNER_REQUEST = "OR", _("Requested by character owner")
         LOST_PERMISSION = "LP", _("Character owner has lost permission")
         MISSING_CORP_TOKEN = "CT", _("Not all corp tokens are recorded in Auth.")
+        REVOKED_IN_GAME = "RG", _("Standing has been revoked in game")
 
     user = models.ForeignKey(
         User, on_delete=models.SET_DEFAULT, default=None, null=True