+++ standingsrequests/tests/test_managers.py
@@ -341,7 +341,10 @@ class TestStandingsRequestValidateRequests(NoSocketsTestCase):
             self.user, 1002, StandingRequest.CHARACTER_CONTACT_TYPE
         )
         StandingRequest.objects.validate_requests()
-        self.assertTrue(StandingRevocation.objects.filter(contact_id=1002).exists())
+        my_revocation = StandingRevocation.objects.get(contact_id=1002)
+        self.assertEqual(
+            my_revocation.reason, StandingRevocation.Reason.LOST_PERMISSION
+        )
 
     def test_create_revocation_if_users_corporation_is_missing_apis(
         self, mock_can_request_corporation_standing
@@ -355,7 +358,10 @@ class TestStandingsRequestValidateRequests(NoSocketsTestCase):
         )
 
         StandingRequest.objects.validate_requests()
-        self.assertTrue(StandingRevocation.objects.filter(contact_id=2001).exists())
+        my_revocation = StandingRevocation.objects.get(contact_id=2001)
+        self.assertEqual(
+            my_revocation.reason, StandingRevocation.Reason.MISSING_CORP_TOKEN
+        )
 
     def test_keep_corp_standing_request_if_all_apis_recorded(
         self, mock_can_request_corporation_standing
@@ -457,7 +463,10 @@ class TestStandingsRevocationManager(NoSocketsTestCase):
 
     def test_add_revocation_new(self):
         my_revocation = StandingRevocation.objects.add_revocation(
-            1001, StandingRevocation.CHARACTER_CONTACT_TYPE, user=self.user_requestor
+            1001,
+            StandingRevocation.CHARACTER_CONTACT_TYPE,
+            user=self.user_requestor,
+            reason=StandingRevocation.Reason.OWNER_REQUEST,
         )
         self.assertIsInstance(my_revocation, StandingRevocation)
 
@@ -482,64 +491,6 @@ class TestStandingsRevocationManager(NoSocketsTestCase):
         )
         self.assertTrue(my_revocation.evaluate_effective_standing())
         self.assertTrue(my_revocation.is_effective)
-
-
-# class TestCharacterAffiliationsManagerAuth(NoSocketsTestCase):
-#     @classmethod
-#     def setUpClass(cls):
-#         super().setUpClass()
-#         cls.user = AuthUtils.create_user("Bruce Wayne")
-
-#     def setUp(self):
-#         EveCharacter.objects.all().delete()
-#         CharacterOwnership.objects.all().delete()
-#         CharacterAffiliation.objects.all().delete()
-#         EveEntity.objects.all().delete()
-
-#     def test_can_update_from_one_character(self):
-#         my_character = create_entity(EveCharacter, 1001)
-#         add_character_to_user(self.user, my_character, is_main=True)
-
-#         CharacterAffiliation.objects.update_evecharacter_relations()
-#         self.assertEqual(CharacterAffiliation.objects.count(), 1)
-#         assoc = CharacterAffiliation.objects.first()
-#         self.assertEqual(assoc.character_id, 1001)
-#         self.assertEqual(assoc.corporation_id, 2001)
-#         self.assertEqual(assoc.main_character_id, 1001)
-#         self.assertEqual(assoc.alliance_id, 3001)
-#         self.assertEqual(
-#             EveEntity.objects.get(id=1001).name, my_character.character_name
-#         )
-
-#     def test_can_handle_no_main(self):
-#         my_character = create_entity(EveCharacter, 1001)
-#         add_character_to_user(self.user, my_character)
-
-#         CharacterAffiliation.objects.update_evecharacter_relations()
-#         self.assertEqual(CharacterAffiliation.objects.count(), 1)
-#         assoc = CharacterAffiliation.objects.first()
-#         self.assertEqual(assoc.character_id, 1001)
-#         self.assertEqual(assoc.corporation_id, 2001)
-#         self.assertIsNone(assoc.main_character_id)
-#         self.assertEqual(assoc.alliance_id, 3001)
-#         self.assertEqual(
-#             EveEntity.objects.get(id=1001).name, my_character.character_name
-#         )
-
-#     def test_can_handle_no_character_without_alliance(self):
-#         my_character = create_entity(EveCharacter, 1004)
-#         add_character_to_user(self.user, my_character)
-
-#         CharacterAffiliation.objects.update_evecharacter_relations()
-#         self.assertEqual(CharacterAffiliation.objects.count(), 1)
-#         assoc = CharacterAffiliation.objects.first()
-#         self.assertEqual(assoc.character_id, 1004)
-#         self.assertEqual(assoc.corporation_id, 2003)
-#         self.assertIsNone(assoc.main_character_id)
-#         self.assertIsNone(assoc.alliance_id)
-#         self.assertEqual(
-#             EveEntity.objects.get(id=1004).name, my_character.character_name
-#         )
 
 
 @patch("standingsrequests.helpers.esi_fetch._esi_client")