+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Please update all contacts after completing the installation. You can do that si
 
 ### Added
 
+- Standing revocations now show reason to managers (#11)
 - Contact sets are now visible on the admin site
 - New button for updating all contacts from the admin site
 - Faction info for characters and corporations on standing pages